[DevEx]: Move i18n.js into `src/plugins` Fixes #1153
