+++ src/main.js
@@ -1,9 +1,9 @@
 import { createApp } from 'vue';
 import { createPinia } from 'pinia';
+import i18n from '@/plugins/i18n';
 import rollbar from '@/plugins/rollbar';
 import vuetify from '@/plugins/vuetify';
 import router from '@/router';
-import i18n from '@/i18n';
 import { createHead } from '@unhead/vue';
 import { initCuttleGlobals } from '_/utils/config-utils';
 import App from '@/App.vue';
